Fixed PMD and Checkstyle errors
